Included Login Endpoint and Salted/Hashed Passwords
First imported pwd.py which contains the necessary functions for salting, hashing, and extracting salts. I included salting and hashing for the customer and employee passwords before storing it in the database in each endpoint. I also added a login endpoint to compare entered password with stored password, both hashed.